blam: add string id mapping
