+++ Reclaimer.Blam/Blam/HaloInfinite/ModuleItem.cs
@@ -104,7 +104,7 @@ namespace Reclaimer.Blam.HaloInfinite
 
                 _cachedTagName = GlobalTagId == -1
                     ? GlobalTagId.ToString()
-                    : TagMapper.TagMappings.TryGetValue(GlobalTagId, out var value)
+                    : StringMapper.TagMappings.TryGetValue(GlobalTagId, out var value)
                         ? value : GlobalTagId.ToString();
 
                 return _cachedTagName;
