+++ Reclaimer/Controls/HaloInfiniteModuleViewer.cs
@@ -67,7 +67,9 @@ namespace Reclaimer.Controls
             InitializeComponent();
 
             if (File.Exists(CommonModuleViewerPlugin.Settings.TagNameFile))
-                TagMapper.LoadTagMap(CommonModuleViewerPlugin.Settings.TagNameFile);
+                StringMapper.LoadTagMap(CommonModuleViewerPlugin.Settings.TagNameFile);
+            if (File.Exists(CommonModuleViewerPlugin.Settings.StringIdFile))
+                StringMapper.LoadStringMap(CommonModuleViewerPlugin.Settings.StringIdFile);
 
             OpenContextItem = new MenuItem { Header = "Open" };
             OpenWithContextItem = new MenuItem { Header = "Open With..." };
